requirements.txt
asgiref==3.9.1
attrs==25.3.0
black==25.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
dj-database-url==3.0.1
django==5.2.6
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
dotenv==0.9.9
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
psycopg==3.2.10
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
pytokens==0.1.10
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
tomli==2.2.1
typing-extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
